added voucher seeder

+++ database/migrations/20200324042952-create-voucher.js
@@ -17,7 +17,7 @@ module.exports = {
         }
       },
       discount: {
-        type: Sequelize.DECIMAL
+        type: Sequelize.SMALLINT
       },
       expiredTime: {
         type: Sequelize.DATE
